more refactoring, animal factory, bug fixes, health decreasing and more!

+++ Savanna.Interfaces/IAnimal.cs
@@ -5,7 +5,7 @@ namespace Savanna.Interfaces
     public interface IAnimal
     {
         AnimalType Type { get; set; }
-        int Health { get; set; }
+        double Health { get; set; }
         char Sign { get; set; }
         int Speed { get; set; }
         int VisionRange { get; set; }
